+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Qobuz App Change Log
 1.2 - unreleased
 ----------------
 
+  * Added support for shuffle.
   * Update for the new Qobuz design.
 
 1.1 - June 17th, 2018